requirements.txt
bcrypt==4.0.1
beanie>=2.0.0
fastapi[all]>=0.120.0
google-generativeai>=0.8.5
motor>=3.7.1
passlib[bcrypt]>=1.7.4
python-dotenv>=1.1.1
python-jose[cryptography]>=3.5.0
uvicorn>=0.38.0
